+++ microsoft/src/cascadia/TerminalApp/AppActionHandlers.cpp
@@ -730,7 +730,7 @@ namespace winrt::TerminalApp::implementation
 
         if (const auto activeTab{ _senderOrFocusedTab(sender) })
         {
-            if (_CurrentWorkspaceLocked())
+            if (_CurrentWorkspaceLocked() || activeTab->IsWorkspaceNodeTab())
             {
                 args.Handled(true);
                 return;
@@ -753,7 +753,7 @@ namespace winrt::TerminalApp::implementation
     {
         if (const auto activeTab{ _senderOrFocusedTab(sender) })
         {
-            if (_CurrentWorkspaceLocked())
+            if (_CurrentWorkspaceLocked() || activeTab->IsWorkspaceNodeTab())
             {
                 args.Handled(true);
                 return;